+++ rstar/src/algorithm/selection_functions.rs
@@ -1,6 +1,6 @@
 use crate::object::PointDistance;
 use crate::object::RTreeObject;
-use crate::{envelope::Envelope, object::Distance};
+use crate::{envelope::Envelope, object::Distance, RTreeNum};
 
 /// Advanced trait to iterate through an r-tree. Usually it should not be required to be implemented.
 ///
@@ -197,7 +197,7 @@ where
 {
     fn should_unpack_parent(&self, parent_envelope: &T::Envelope) -> bool {
         let envelope_distance = parent_envelope.distance_2(&self.circle_origin);
-        envelope_distance <= self.squared_max_distance
+        envelope_distance.ord() <= self.squared_max_distance.ord()
     }
 
     fn should_unpack_leaf(&self, leaf: &T) -> bool {